git-svn-id: https://svn.sdss.org/repo/survey_software/sdss_python_module/trunk@8812 acfa968b-e9c4-4993-a9e6-846a59bbf143

+++ dbclasses/plate_utils.py
@@ -234,7 +234,7 @@ def rearrangeSets(plate, mode='complete', scope='all', force=False,
         raise exceptions.TotoroError('scope={0} is invalid'.format(scope))
 
     # Rejects invalid exposures:
-    exposures = [exp for exp in exposures if exp.valid]
+    exposures = [exp for exp in exposures if exp.isPlateDBValid()]
 
     # Does some logging.
     logMode('plate_id={0}: rearranging sets, mode=\'{1}\', '
@@ -250,8 +250,10 @@ def rearrangeSets(plate, mode='complete', scope='all', force=False,
         # and triggers a plate update.
         with session.begin(subtransactions=True):
             for exposure in exposures:
-                session.delete(exposure.mangadbExposure[0].set)
+                if exposure.mangadbExposure[0].set_pk is not None:
+                    session.delete(exposure.mangadbExposure[0].set)
                 exposure.mangadbExposure[0].set_pk = None
+                exposure.mangadbExposure[0].exposure_status_pk = None
 
         plate.sets = []
         updatePlate(plate, rearrangeIncomplete=False)
